\name{CVSCIR-package}
\alias{CVSCIR-package}
\alias{CVSCIR}
\docType{package}
\title{
\packageTitle{CVSCIR}
}
\description{
\packageDescription{CVSCIR}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{CVSCIR}

Maintainer: \packageMaintainer{CVSCIR}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = CVSCIR)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{CVSCIR}}
%\section{Documentation Index}{\packageIndices{CVSCIR}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
